Melhoras no sequencial

+++ src/utils.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 #include "estruturas.h"
 
-void inicializaRandom(int seed);
+void inicializaRandom(unsigned long long seed);
 double randomDouble();
 node* selecaoRoleta(listaLigada *filhos, int qtaFilhos, double alfa, double beta);
